requirements.txt
google-generativeai 
langchain 
langgraph 
python-dotenv 
streamlit
langchain-google-genai